[Refactor]: Consolidate GQA‑BWD perf script into test file
- Remove standalone perf_example_gqa_bwd.py
- Move perf‑related logic into test_gqa_bwd.py
- Update design.md
